requirements.txt
django==3.1.5
djangorestframework==3.12.4
drf-yasg==1.20.0
djangorestframework-simplejwt
factory-boy